Move ContextMenu out of editor.rs Rename ContextMenu to CodeContextMenu
